passed target in validation set as int, since it should always be labeled data

+++ nnunetv2/training/nnUNetTrainer/semiSupervisedTrainer.py
@@ -1201,7 +1201,8 @@ class semiSupervisednnUNetTrainer(nnUNetTrainer):
         else:
             mask = None
 
-        tp, fp, fn, _ = get_tp_fp_fn_tn(predicted_segmentation_onehot, target, axes=axes, mask=mask)
+        # since all the validation set should be labeled, we can safely turn it into an int for this function
+        tp, fp, fn, _ = get_tp_fp_fn_tn(predicted_segmentation_onehot, target.long(), axes=axes, mask=mask)
 
         tp_hard = tp.detach().cpu().numpy()
         fp_hard = fp.detach().cpu().numpy()
